requirements.txt
requests
beautifulsoup4
pyyaml
requests-html
pyppeteer
lxml_html_clean
selenium
webdriver-manager